\name{hbiclust-package}
\alias{hbiclust-package}
\alias{hbiclust}
\docType{package}
\title{
\packageTitle{hbiclust}
}
\description{
\packageDescription{hbiclust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hbiclust}
\packageIndices{hbiclust}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hbiclust}

Maintainer: \packageMaintainer{hbiclust}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
